break state validation into a `utils` module
